requirements.txt
Flask
streamlit
requests
python-dotenv
beautifulsoup4
langchain-groq